refactor: Extract auth into standalone smg-auth workspace crate
- Created auth/Cargo.toml with lib name smg_auth
- Moved source files from src/auth to auth/src
- Updated imports from super:: to crate::
- Added RequestId type to auth crate for audit logging
- Exported AuthMethod from middleware module
- Updated root Cargo.toml workspace members
- Re-exported auth crate as smg::auth in main lib
