+++ auth/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "auth"
+version = "0.1.0"
+edition = "2021"
+description = "Authentication and authorization for control plane APIs"
+license = "Apache-2.0"
+repository = "https://github.com/lightseekorg/smg"
+keywords = ["auth", "jwt", "oidc", "api-key", "middleware"]
+categories = ["authentication", "web-programming"]
+
+[lib]
+name = "smg_auth"
+
+[dependencies]
+axum = { version = "0.8.1", features = ["macros"] }
+chrono = { version = "0.4", features = ["serde"] }
+http = "1.2"
+jsonwebtoken = "9.3"
+lru = "0.16.2"
+parking_lot = "0.12.4"
+reqwest = { version = "0.12.8", default-features = false, features = ["rustls-tls", "json"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+sha2 = "0.10"
+subtle = "2.6"
+thiserror = "2.0.12"
+tracing = "0.1"
+url = "2.5"
+
+[dev-dependencies]
+tokio = { version = "1.42.0", features = ["rt-multi-thread", "macros"] }
