+++ auth/src/lib.rs
@@ -16,5 +16,13 @@ pub use audit::{AuditEvent, AuditLogger, AuditOutcome};
 pub use config::{ApiKeyEntry, ControlPlaneAuthConfig, JwtConfig, Role};
 pub use jwt::{JwtValidator, JwtValidatorError};
 pub use middleware::{
-    control_plane_auth_middleware, ControlPlaneAuthState, Principal, PrincipalExt,
+    control_plane_auth_middleware, AuthMethod, ControlPlaneAuthState, Principal, PrincipalExt,
 };
+
+/// Request ID for correlation in audit logs.
+///
+/// This type can be added to request extensions to provide request IDs
+/// for audit logging. The middleware will extract this from request
+/// extensions if present.
+#[derive(Debug, Clone)]
+pub struct RequestId(pub String);
